% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedHypTestUtilities.R
\name{alpha_weak}
\alias{alpha_weak}
\title{alpha_weak}
\usage{
alpha_weak(max_n_segments, alpha_total, alpha_strong)
}
\arguments{
\item{max_n_segments}{Maximum number of segments to be run before termination}

\item{alpha_total}{Desired total Type 1 error probability across the entire
study}

\item{alpha_strong}{For decision rule: If p-observed for a segment is less
than alpha strong, stop and reject Ho.}
}
\value{
Numeric. Value of alpha weak.
}
\description{
\code{alpha_weak}  returns the value of alpha_weak which
will produce the correct overall Type 1 error probability (i.e. alpha_total)
for a segmented hypothesis testing design, given values for maximum
number of segments, alpha total and alpha strong. This function wraps the
alpha_weak_computation method of the SegmentedResearcher class. If alpha
strong has been specified as zero (boundary case) alpha weak has an
analytical solution alpha_total^(1/max_n_segments). Otherwise alpha weak is
found via numerical search using function pracma::fzero. See Miller and
Ulrich, 2020 for discussion.
}
