% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{compute_power}
\alias{compute_power}
\title{compute_power}
\usage{
compute_power(
  stat_procedure,
  alpha_one_tailed,
  effect_size,
  sample_n,
  critical_value = NULL
)
}
\arguments{
\item{stat_procedure}{Instance for method dispatch. Member of StatProcedureBase family}

\item{alpha_one_tailed}{Alpha level for one-tailed test}

\item{effect_size}{Presumed true effect size}

\item{sample_n}{Group sizes. Must be a vector for tests with multiple groups}

\item{critical_value}{Optional to speed computation. If omitted, this value is computed in the method}
}
\value{
Computed statistical power (probability of rejecting H0 when the effect is present)
}
\description{
Abstract method defined in StatProcedureBase. Should be overridden by all children
}
\details{
\code{compute_power} computes the statistical power of a statistical test
}
