% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureTwoSampleT.R
\name{compute_power.TwoSampleT}
\alias{compute_power.TwoSampleT}
\title{compute_power.TwoSampleT}
\usage{
\method{compute_power}{TwoSampleT}(
  stat_procedure,
  alpha_one_tailed,
  effect_size,
  sample_n,
  critical_value = NULL
)
}
\arguments{
\item{stat_procedure}{TwoSampleT instance for method dispatch}

\item{alpha_one_tailed}{Alpha level for one-tailed test}

\item{effect_size}{Presumed true effect size on Cohen's d scale}

\item{sample_n}{Group sizes. Must be a numeric vector of length 2}

\item{critical_value}{Optional to speed computation. If omitted, this value is computed in the method}
}
\value{
Computed statistical power (probability of rejecting H0 when the effect is present)
}
\description{
Implementation of compute_power method for class TwoSampleT
}
\details{
\code{TwoSampleT} computes the statistical power of a described two sample
t-test as the probability density of t-critical in the t-distribution for a given effect size
}
