% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{divide_total_sample}
\alias{divide_total_sample}
\title{divide_total_sample}
\usage{
divide_total_sample(stat_procedure, n_per_segment)
}
\arguments{
\item{stat_procedure}{Instance for method dispatch. Member of StatProcedureBase family}

\item{n_per_segment}{Total n per segment}
}
\value{
N per group for the associated statistical procedure object
}
\description{
Abstract method defined in StatProcedureBase. Should be overridden by all
children
}
\details{
\code{divide_total_sample} returns n per group given total n
}
