% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedHypTestUtilities.R
\name{n_for_power}
\alias{n_for_power}
\title{n_for_power}
\usage{
n_for_power(
  target_power,
  stat_procedure_name,
  effect_size,
  max_n_segments = 3,
  alpha_total = 0.5,
  alpha_strong = 0.025,
  alpha_weak = NA
)
}
\arguments{
\item{target_power}{Desired power level (probability of detecting a false H0
when present) across all segments of the study}

\item{stat_procedure_name}{Character code for hypothesis testing procedure to
be used in the study. Currently must be one of {'1t', '2t', 'r', '1z',
'2z'}}

\item{effect_size}{Size of the true effect when it is present.
For tests of correlation, this is the true correlation rho.
For tests of means (1t, 2t, 1z, 2z), this Cohen's d.}

\item{max_n_segments}{Maximum number of segments to be run before termination}

\item{alpha_total}{Desired total Type 1 error probability across the entire
study}

\item{alpha_strong}{For decision rule: If p-observed for a segment is less
than alpha_strong, stop and reject Ho.}

\item{alpha_weak}{For decision rule: If p-observed for a segment is greater
than alpha weak, stop and fail to reject Ho. This argument is not normally provided,
in which case the correct value is computed in the method. If the argument is
provided, it is used without being checked (for speed) so it is the user's
responsibility to ensure that it is the correct value.}
}
\value{
Returns the number
  of subjects per segment that yields the target power. The result is not
  necessarily a whole number, in which case you must use the next
  smaller or larger integer and get a little lower or higher power.
}
\description{
\code{n_for_power}  returns the number of subjects
required to achieve a target statistical power for a segmented
hypothesis testing design, given the other design descriptors (see Arguments,
below). This method wraps n_for_power_computation to prevent the end user from
having to create his/her own component class instances. The result is
determined by numerical search using pracma::fminbnd. The function minimised
computes the difference between the power obtained by a given value of n and
the target power. The power obtained for any given value of n is computed
numerically. See n_for_power_computation in this file and
seght::run_scenario.SegmentedHypTestEngine for computation logic.
}
