% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{noncentrality}
\alias{noncentrality}
\title{noncentrality}
\usage{
noncentrality(stat_procedure, sample_n, effect_size)
}
\arguments{
\item{stat_procedure}{Instance for method dispatch. Member of StatProcedureBase family}

\item{sample_n}{Group sizes. Must be a vector for tests with multiple groups}

\item{effect_size}{Presumed true effect size}
}
\value{
Computed noncentrality parameter
}
\description{
Abstract method defined in StatProcedureBase. Should be overridden by all children
}
\details{
\code{noncentrality} computes the noncentrality parameter for a given n and effect size
}
