% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedurePearsonR.R
\name{PearsonR}
\alias{PearsonR}
\title{PearsonR Constructor}
\usage{
PearsonR(min_sample_n = 5, display_abbrev = "r", display_name = "Pearson r")
}
\arguments{
\item{min_sample_n}{Smallest n allowed for this test. Supplied in constructor.}

\item{display_abbrev}{Two-character abbreviation for this test. Developers see stat_procedure_factory.SegmentedHypTester}

\item{display_name}{String value useful for generating tidy outputs for this test}
}
\value{
Validated PearsonR instance
}
\description{
PearsonR is a child of StatProcedureBase, representing a Pearson Product Moment correlation
}
\details{
\code{PearsonR}  is the public-facing PearsonR constructor. It creates and returns a class instance
with appropriately initialised values for all fields (min_sample_n, display_abbrev, and
display_name). It requires no parameters.
}
