% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedHypTestUtilities.R
\name{search_kmax}
\alias{search_kmax}
\title{search_kmax}
\usage{
search_kmax(
  alpha_total,
  alpha_strong,
  stat_procedure_name,
  target_power,
  effect_size,
  base_rate = 1
)
}
\arguments{
\item{alpha_total}{Desired total Type 1 error probability across the entire
study}

\item{alpha_strong}{For decision rule: If p-observed for a segment is less
than alpha strong, stop and reject Ho.}

\item{stat_procedure_name}{Character code for hypothesis testing procedure to
be used in the experiment. Currently must be one of {'1t', '2t', 'r', '1z',
'2z'}}

\item{target_power}{Desired power level (probability of detecting a false H0
when present) across all segments of the study}

\item{effect_size}{Size of the true effect when it is present.
For tests of correlation, this is the true correlation rho.
For tests of means (1t, 2t, 1z, 2z), this Cohen's d.}

\item{base_rate}{Base rate of the true effect_size. Must be between 0 and 1.
If this value is less than 1, the remaining probability is allocated to
effect size 0 (i.e. H0 is true)}
}
\value{
Returns a data frame with one row for each value of kmax between 2
  and 10, and columns: \itemize{ \item kmax: Maximum number of segments in
  the study \item alpha_weak: For decision rule: If p-observed for a segment
  is greater than alpha weak, stop and fail to reject Ho. \item n per segment:
  Number of subjects per segment needed to achieve the target power. \item expected n subjects: Expected (mean)
  number of subjects required across infinite iterations of the experiment
  \item expected n segments: Expected (mean) number of segments that will be
  performed prior to termination (with any outcome) across infinite
  iterations of the experiment. }
}
\description{
\code{search_kmax} computes a set  of performance
descriptors (see Values, below) for a specified study scenario for each
value of kmax (maximum number of segments) between 2 and 10. Returned values
are based on the n_per_segment required to achieve the provided target statistical power.
}
