% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedHypTestUtilities.R
\name{segmented_hyp_test_outcomes}
\alias{segmented_hyp_test_outcomes}
\title{segmented_hyp_test_outcomes}
\usage{
segmented_hyp_test_outcomes(
  max_n_segments,
  n_per_segment,
  alpha_total,
  alpha_strong,
  stat_procedure_name,
  effect_sizes,
  base_rates = 1
)
}
\arguments{
\item{max_n_segments}{Maximum number of segments to be run before termination}

\item{n_per_segment}{Number of subjects to be tested in each segment}

\item{alpha_total}{Desired total Type 1 error probability across the entire
experiment}

\item{alpha_strong}{For decision rule: If p-observed for a segment is less
than alpha strong, stop and reject Ho.}

\item{stat_procedure_name}{Character code for hypothesis testing procedure to
be used in the experiment. Currently must be one of {'1t', '2t', 'r', '1z',
'2z'}}

\item{effect_sizes}{Value, or vector of values, representing the true effect when it is present.
For tests of correlation, this is the true correlation rho.
For tests of means (1t, 2t, 1z, 2z), this Cohen's d.
Any effect size of zero is logically Ho = true.}

\item{base_rates}{Base rate or vector of base rates giving the probability of each effect size.
Each base_rate value matches the corresponding element of the effect_sizes argument, above.
Each base rate must be between 0 and 1, and their total should be less than or equal to 1.
If the total of base_rates is less than 1, the
remaining probability is allocated to effect size 0 (i.e. H0 is true)}
}
\value{
Returns a list (collection of named items) containing all of the method
  input values (for reference), plus: \itemize{ \item exp_n_sub: Expected (mean
  across infinite iterations) number of subjects used in the study \item
  avg_power: Power level of the study on average across all cases where Ho is false
  (weighted average according to base rates).
  \item pr_reject_by_segment: Probablility of rejecting H0 at each segment
  \item pr_ftr_by_segment: Probability of failing to reject H0 at each
  segment }
}
\description{
\code{segmented_hyp_test_outcomes}  wraps
seght::run_scenario.SegmentedHypTestEngine to numerically determine the
probabilities of all outcomes (reject, fail to reject, or continue) for each
potential segment of a segmented hypothesis testing design, along with a
variety of useful summary descriptors of expected study performance.
}
