% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedHypTestEngine.R
\name{SegmentedHypTestEngine}
\alias{SegmentedHypTestEngine}
\title{SegmentedHypTestEngine}
\usage{
SegmentedHypTestEngine()
}
\value{
Validated SegmentedHypTestEngine instance
}
\description{
\code{SegmentedHypTestEngine}  is the public-facing SegmentedHypTestEngine
constructor. It accepts no parameters. This object exposes the primary methods
for numerically determing the expected outcomes of a segmented hypothesis testing
study. See run_scenario.SegmentedHypTestEngine for an example.
}
