% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatProcedureBase.R
\name{StatProcedureBase}
\alias{StatProcedureBase}
\title{StatProcedureBase Placeholder Ctor}
\usage{
StatProcedureBase(min_sample_n, display_abbrev, display_name)
}
\arguments{
\item{min_sample_n}{Smallest n allowed for this test. Supplied in constructor.}

\item{display_abbrev}{Two-character abbreviation for this test. Developers see stat_procedure_factory.SegmentedHypTester}

\item{display_name}{String value useful for generating tidy outputs for this test}
}
\description{
StatProcedureBase is the abstract base class for all stat procedure classes.
}
\details{
\code{StatProcedureBase}  should not be called directly. Descend all child classes
from StatProcedureBase by assigning the class attribute in the child constructor.
For example, the low-level constructor for class OneSampleT contains the statement:

\code{attr(new_one_sample_t, "class") <- c("OneSampleT", "StatProcedureBase")}
}
